requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
discord==2.3.2
discord.py==2.4.0
distro==1.9.0
frozenlist==1.4.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.66.0
grpcio-status==1.62.3
gspread==6.1.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
multidict==6.0.5
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.35.10
pillow==10.4.0
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.4
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
responses==0.25.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
yarl==1.9.4
